Rename cuda specific files
polyhedral/cuda/cuda_mapping_types.h -> polyhedral/cuda/mapping_types.h
polyhedral/cuda/cuda_tighten_launch_bounds.h -> polyhedral/cuda/tighten_launch_bounds.h
