+++ include/tc/core/polyhedral/schedule_transforms.h
@@ -21,7 +21,7 @@
 #include <unordered_set>
 #include <vector>
 
-#include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
+#include "tc/core/polyhedral/cuda/mapping_types.h"
 #include "tc/core/polyhedral/functional.h"
 #include "tc/core/polyhedral/options.h"
 #include "tc/core/polyhedral/schedule_tree.h"
