+++ include/tc/core/polyhedral/schedule_tree_elem.h
@@ -22,7 +22,7 @@
 
 #include "tc/external/isl.h"
 
-#include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
+#include "tc/core/polyhedral/cuda/mapping_types.h"
 
 namespace tc {
 namespace polyhedral {
