+++ src/autotuner/genetic_tuning_harness.cc
@@ -32,7 +32,7 @@
 #include "tc/core/cuda/cuda_tc_executor.h"
 #include "tc/core/execution_engine.h"
 #include "tc/core/flags.h"
-#include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
+#include "tc/core/polyhedral/cuda/mapping_types.h"
 #include "tc/core/scope_guard.h"
 #include "tc/core/utils/math.h"
