+++ src/core/polyhedral/cuda/codegen.cc
@@ -26,8 +26,8 @@
 #include "tc/core/libraries.h"
 #include "tc/core/polyhedral/codegen.h"
 #include "tc/core/polyhedral/cuda/codegen.h"
-#include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
 #include "tc/core/polyhedral/memory_promotion.h"
+#include "tc/core/polyhedral/cuda/mapping_types.h"
 #include "tc/core/polyhedral/schedule_isl_conversion.h"
 #include "tc/core/polyhedral/schedule_transforms.h"
