+++ src/core/polyhedral/cuda/mapped_scop.cc
@@ -27,11 +27,11 @@
 #include "tc/core/gpu.h"
 #include "tc/core/libraries.h"
 #include "tc/core/polyhedral/cuda/codegen.h"
-#include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
-#include "tc/core/polyhedral/cuda/cuda_tighten_launch_bounds.h"
+#include "tc/core/polyhedral/cuda/mapping_types.h"
+#include "tc/core/polyhedral/cuda/memory_promotion_heuristic.h"
+#include "tc/core/polyhedral/cuda/tighten_launch_bounds.h"
 #include "tc/core/polyhedral/exceptions.h"
 #include "tc/core/polyhedral/functional.h"
-#include "tc/core/polyhedral/cuda/memory_promotion_heuristic.h"
 #include "tc/core/polyhedral/schedule_transforms.h"
 #include "tc/core/polyhedral/schedule_tree_matcher.h"
 #include "tc/core/polyhedral/scop.h"
