+++ test/test_mapper.cc
@@ -27,9 +27,9 @@
 #include "tc/core/cuda/cuda_mapping_options.h"
 #include "tc/core/libraries.h"
 #include "tc/core/polyhedral/cuda/codegen.h"
-#include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
-#include "tc/core/polyhedral/functional.h"
 #include "tc/core/polyhedral/cuda/mapped_scop.h"
+#include "tc/core/polyhedral/cuda/mapping_types.h"
+#include "tc/core/polyhedral/functional.h"
 #include "tc/core/polyhedral/schedule_isl_conversion.h"
 #include "tc/core/polyhedral/schedule_transforms.h"
 #include "tc/core/polyhedral/schedule_tree.h"
